feat(ui): replace dropdown with lock & deploy buttons
