+++ services/frontend-service/src/ui/components/button/button.scss
@@ -32,33 +32,23 @@ Copyright freiheit.com*/
     }
 }
 
-.expand-button {
-    margin-right: 20px;
-    border-width: thin;
-    border-radius: 14px;
-    border-color: blue;
-    border-style: groove;
-    .button-popup-deploy {
-        &:disabled {
-            color: grey;
-        }
+.deploy-lock-buttons {
+    height: 36px;
+    margin-right: 10px;
+
+    .mdc-button {
+        margin-right: 10px;
+        border-width: thin;
+        border-radius: 12px 0 0 12px;
+        border-color: blue;
+        border-style: groove;
+        width: 205px;
     }
 
-    .first-two {
-        height: 36px;
-        .button-main {
-            border-radius: 12px 0 0 12px;
-            &:disabled {
-                color: white;
-            }
-        }
-        .button-expand {
-            padding: 0;
-            min-width: 36px;
-            .dropdown-arrow {
-                margin-bottom: 40%; // to center the arrow
-                font-size: xx-large;
-            }
+    .button-main {
+        width: 170px;
+        &:disabled {
+            color: white;
         }
     }
 }
